Multiple minor fixes. Moved InMemoryAppender-subscription functionality to Singleton, split InMemoryAppender and LogConfigBuilder
