+++ platform-sdk/consensus-otter-docker-app/src/testFixtures/java/org/hiero/consensus/otter/docker/app/DockerMain.java
@@ -10,7 +10,7 @@ import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
-import org.hiero.otter.fixtures.logging.LogConfigBuilder;
+import org.hiero.consensus.otter.docker.app.logging.DockerLogConfigBuilder;
 
 /**
  * Main entry point for the Docker container application.
@@ -71,7 +71,7 @@ public final class DockerMain {
      * @throws InterruptedException if the server is interrupted while waiting for termination
      */
     public static void main(final String[] args) throws IOException, InterruptedException {
-        LogConfigBuilder.configureNode(Path.of(""));
+        DockerLogConfigBuilder.configure(Path.of(""), null);
         new DockerMain(createDispatchExecutor(), Executors.newCachedThreadPool()).startGrpcServer();
     }
