+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/ProtobufConverter.java
@@ -652,7 +652,7 @@ public class ProtobufConverter {
                 sourceLog.getLoggerName(),
                 sourceLog.getThread(),
                 MarkerManager.getMarker(sourceLog.getMarker()),
-                NodeId.newBuilder().id(sourceLog.getNodeId()).build());
+                sourceLog.getNodeId() < 0? null : NodeId.newBuilder().id(sourceLog.getNodeId()).build());
     }
 
     /**
