+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/assertions/SingleNodeLogResultAssert.java
@@ -117,7 +117,8 @@ public class SingleNodeLogResultAssert extends AbstractAssert<SingleNodeLogResul
         logStatements.append("\n****************\n");
         logStatements.append(" ->  Log messages found:\n");
         logStatements.append("****************\n");
-        logs.forEach(log -> logStatements.append(log.toString()));
+        // logStatements is used as a formatter String, thus we must escape '%'
+        logs.forEach(log -> logStatements.append(log.toString().replace("%", "%%")));
         logStatements.append("****************\n");
 
         failWithMessage(logStatements.toString());
