+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/result/NodeResultsCollector.java
@@ -8,15 +8,20 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Queue;
+import java.util.Set;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.CopyOnWriteArrayList;
+import org.apache.logging.log4j.Marker;
 import org.hiero.consensus.model.hashgraph.ConsensusRound;
 import org.hiero.consensus.model.status.PlatformStatus;
+import org.hiero.otter.fixtures.logging.StructuredLog;
 import org.hiero.otter.fixtures.result.ConsensusRoundSubscriber;
+import org.hiero.otter.fixtures.result.LogSubscriber;
 import org.hiero.otter.fixtures.result.MarkerFileSubscriber;
 import org.hiero.otter.fixtures.result.MarkerFilesStatus;
 import org.hiero.otter.fixtures.result.PlatformStatusSubscriber;
 import org.hiero.otter.fixtures.result.SingleNodeConsensusResult;
+import org.hiero.otter.fixtures.result.SingleNodeLogResult;
 import org.hiero.otter.fixtures.result.SingleNodePlatformStatusResult;
 import org.hiero.otter.fixtures.result.SubscriberAction;
 
@@ -30,6 +35,8 @@ public class NodeResultsCollector {
     private final List<ConsensusRoundSubscriber> consensusRoundSubscribers = new CopyOnWriteArrayList<>();
     private final List<PlatformStatus> platformStatuses = new ArrayList<>();
     private final List<PlatformStatusSubscriber> platformStatusSubscribers = new CopyOnWriteArrayList<>();
+    private final List<StructuredLog> logEntries = new ArrayList<>();
+    private final List<LogSubscriber> logSubscribers = new CopyOnWriteArrayList<>();
     private final List<MarkerFileSubscriber> markerFileSubscribers = new CopyOnWriteArrayList<>();
 
     // This class may be used in a multi-threaded context, so we use volatile to ensure visibility of state changes
@@ -84,12 +91,26 @@ public class NodeResultsCollector {
     }
 
     /**
+     * Adds a log entry to the list of collected logs.
+     *
+     * @param logEntry the {@link StructuredLog} to add
+     */
+    public void addLogEntry(@NonNull final StructuredLog logEntry) {
+        requireNonNull(logEntry);
+        if (!destroyed) {
+            logEntries.add(logEntry);
+            logSubscribers.removeIf(
+                    subscriber -> subscriber.onLogEntry(logEntry) == SubscriberAction.UNSUBSCRIBE);
+        }
+    }
+
+    /**
      * Returns a {@link SingleNodeConsensusResult} of the current state.
      *
      * @return the {@link SingleNodeConsensusResult}
      */
     @NonNull
-    public SingleNodeConsensusResult getConsensusResult() {
+    public SingleNodeConsensusResult newConsensusResult() {
         return new SingleNodeConsensusResultImpl(this);
     }
 
@@ -103,6 +124,15 @@ public class NodeResultsCollector {
     public List<ConsensusRound> currentConsensusRounds(final int startIndex) {
         final List<ConsensusRound> copy = List.copyOf(consensusRounds);
         return copy.subList(startIndex, copy.size());
+    }
+
+    /**
+     * Returns the number of consensus rounds created by this node.
+     *
+     * @return the count of consensus rounds
+     */
+    public int currentConsensusRoundsCount() {
+        return consensusRounds.size();
     }
 
     /**
@@ -121,7 +151,7 @@ public class NodeResultsCollector {
      * @return the {@link SingleNodePlatformStatusResult}
      */
     @NonNull
-    public SingleNodePlatformStatusResult getStatusProgression() {
+    public SingleNodePlatformStatusResult newStatusProgression() {
         return new SingleNodePlatformStatusResultImpl(this);
     }
 
@@ -131,9 +161,19 @@ public class NodeResultsCollector {
      * @param startIndex the index to start from
      * @return the list of platform statuses
      */
+    @NonNull
     public List<PlatformStatus> currentStatusProgression(final int startIndex) {
         final List<PlatformStatus> copy = List.copyOf(platformStatuses);
         return copy.subList(startIndex, copy.size());
+    }
+
+    /**
+     * Returns the number of platform statuses created by this node.
+     *
+     * @return the count of platform statuses
+     */
+    public int currentStatusProgressionCount() {
+        return platformStatuses.size();
     }
 
     /**
@@ -144,6 +184,50 @@ public class NodeResultsCollector {
     public void subscribePlatformStatusSubscriber(@NonNull final PlatformStatusSubscriber subscriber) {
         requireNonNull(subscriber);
         platformStatusSubscribers.add(subscriber);
+    }
+
+    /**
+     * Returns a new {@link SingleNodeLogResult} for the node.
+     *
+     * @return the new {@link SingleNodeLogResult}
+     */
+    @NonNull
+    public SingleNodeLogResult newLogResult() {
+        return new SingleNodeLogResultImpl(this, Set.of());
+    }
+
+    /**
+     * Returns all the log entries created at the moment of invocation, starting with and including the provided index.
+     *
+     * @param startIndex the index to start from
+     * @param suppressedLogMarkers the set of {@link Marker} that should be ignored in the logs
+     * @return the list of log entries
+     */
+    @NonNull
+    public List<StructuredLog> currentLogEntries(final long startIndex, @NonNull final Set<Marker> suppressedLogMarkers) {
+        return logEntries.stream()
+                .skip(startIndex)
+                .filter(logEntry -> logEntry.marker() == null || !suppressedLogMarkers.contains(logEntry.marker()))
+                .toList();
+    }
+
+    /**
+     * Returns the number of log entries created by this node.
+     *
+     * @return the count of log entries
+     */
+    public int currentLogEntriesCount() {
+        return logEntries.size();
+    }
+
+    /**
+     * Subscribes to log events for the node.
+     *
+     * @param subscriber the subscriber that will receive log events
+     */
+    public void subscribeLogSubscriber(@NonNull final LogSubscriber subscriber) {
+        requireNonNull(subscriber);
+        logSubscribers.add(subscriber);
     }
 
     /**
