+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/result/SingleNodeConsensusResultImpl.java
@@ -18,7 +18,7 @@ public class SingleNodeConsensusResultImpl implements SingleNodeConsensusResult 
 
     private final NodeResultsCollector collector;
 
-    // This class may be used in a multi-threaded context, so we use volatile to ensure visibility of state changes
+    // This class may be used in a multithreaded context, so we use volatile to ensure visibility of state changes
     private volatile int startIndex = 0;
 
     /**
@@ -74,6 +74,6 @@ public class SingleNodeConsensusResultImpl implements SingleNodeConsensusResult 
      */
     @Override
     public void clear() {
-        startIndex = collector.currentConsensusRounds(0).size();
+        startIndex = collector.currentConsensusRoundsCount();
     }
 }
