+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/result/SingleNodeLogResultImpl.java
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.otter.fixtures.internal.result;
 
-import static java.util.Collections.unmodifiableSet;
 import static java.util.Objects.requireNonNull;
 
 import com.hedera.hapi.platform.state.NodeId;
@@ -9,11 +8,9 @@ import com.swirlds.logging.legacy.LogMarker;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import org.apache.logging.log4j.Marker;
 import org.hiero.otter.fixtures.logging.StructuredLog;
-import org.hiero.otter.fixtures.logging.internal.InMemoryAppender;
 import org.hiero.otter.fixtures.result.LogSubscriber;
 import org.hiero.otter.fixtures.result.SingleNodeLogResult;
 import org.hiero.otter.fixtures.result.SubscriberAction;
@@ -23,21 +20,21 @@ import org.hiero.otter.fixtures.result.SubscriberAction;
  */
 public class SingleNodeLogResultImpl implements SingleNodeLogResult {
 
-    private final NodeId nodeId;
+    private final NodeResultsCollector collector;
     private final Set<Marker> suppressedLogMarkers;
 
-    // This class may be used in a multi-threaded context, so we use volatile to ensure visibility of state changes
-    private volatile int startIndex = 0;
+    // This class may be used in a multithreaded context, so we use volatile to ensure visibility of state changes
+    private volatile long startIndex = 0;
 
     /**
      * Creates a new instance of {@link SingleNodeLogResultImpl}.
      *
-     * @param nodeId the {@link NodeId} of the node for which logs are collected
+     * @param collector the {@link NodeResultsCollector} that collects the results
      * @param suppressedLogMarkers the set of {@link Marker} that should be ignored in the logs
      */
-    public SingleNodeLogResultImpl(@NonNull final NodeId nodeId, @NonNull final Set<Marker> suppressedLogMarkers) {
-        this.nodeId = requireNonNull(nodeId);
-        this.suppressedLogMarkers = unmodifiableSet(requireNonNull(suppressedLogMarkers));
+    public SingleNodeLogResultImpl(@NonNull final NodeResultsCollector collector, @NonNull final Set<Marker> suppressedLogMarkers) {
+        this.collector = requireNonNull(collector);
+        this.suppressedLogMarkers = Set.copyOf(suppressedLogMarkers);
     }
 
     /**
@@ -46,7 +43,7 @@ public class SingleNodeLogResultImpl implements SingleNodeLogResult {
     @Override
     @NonNull
     public NodeId nodeId() {
-        return nodeId;
+        return collector.nodeId();
     }
 
     /**
@@ -55,10 +52,7 @@ public class SingleNodeLogResultImpl implements SingleNodeLogResult {
     @Override
     @NonNull
     public List<StructuredLog> logs() {
-        return InMemoryAppender.getLogs(nodeId).stream()
-                .skip(startIndex)
-                .filter(logEntry -> logEntry.marker() == null || !suppressedLogMarkers.contains(logEntry.marker()))
-                .toList();
+        return collector.currentLogEntries(startIndex, suppressedLogMarkers);
     }
 
     /**
@@ -72,7 +66,7 @@ public class SingleNodeLogResultImpl implements SingleNodeLogResult {
         final Set<Marker> markers = new HashSet<>(suppressedLogMarkers);
         markers.add(marker.getMarker());
 
-        return new SingleNodeLogResultImpl(nodeId, markers);
+        return new SingleNodeLogResultImpl(collector, markers);
     }
 
     /**
@@ -81,12 +75,12 @@ public class SingleNodeLogResultImpl implements SingleNodeLogResult {
     @Override
     public void subscribe(@NonNull final LogSubscriber subscriber) {
         final LogSubscriber wrapper = logEntry -> {
-            if (Objects.equals(this.nodeId, logEntry.nodeId()) && !suppressedLogMarkers.contains(logEntry.marker())) {
+            if (!suppressedLogMarkers.contains(logEntry.marker())) {
                 return subscriber.onLogEntry(logEntry);
             }
             return SubscriberAction.CONTINUE;
         };
-        InMemoryAppender.subscribe(wrapper);
+        collector.subscribeLogSubscriber(wrapper);
     }
 
     /**
@@ -94,6 +88,6 @@ public class SingleNodeLogResultImpl implements SingleNodeLogResult {
      */
     @Override
     public void clear() {
-        startIndex = InMemoryAppender.getLogs(nodeId).size();
+        startIndex = collector.currentLogEntriesCount();
     }
 }
