+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/turtle/TurtleNode.java
@@ -11,6 +11,8 @@ import static org.hiero.otter.fixtures.internal.AbstractNode.LifeCycle.DESTROYED
 import static org.hiero.otter.fixtures.internal.AbstractNode.LifeCycle.INIT;
 import static org.hiero.otter.fixtures.internal.AbstractNode.LifeCycle.RUNNING;
 import static org.hiero.otter.fixtures.internal.AbstractNode.LifeCycle.SHUTDOWN;
+import static org.hiero.otter.fixtures.result.SubscriberAction.CONTINUE;
+import static org.hiero.otter.fixtures.result.SubscriberAction.UNSUBSCRIBE;
 import static org.hiero.otter.fixtures.turtle.TurtleTestEnvironment.APP_NAME;
 import static org.hiero.otter.fixtures.turtle.TurtleTestEnvironment.SWIRLD_NAME;
 
@@ -44,7 +46,7 @@ import edu.umd.cs.findbugs.annotations.Nullable;
 import java.nio.file.Path;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.Set;
+import java.util.Objects;
 import org.apache.logging.log4j.ThreadContext;
 import org.hiero.consensus.model.node.KeysAndCerts;
 import org.hiero.consensus.model.status.PlatformStatus;
@@ -58,9 +60,9 @@ import org.hiero.otter.fixtures.app.OtterApp;
 import org.hiero.otter.fixtures.app.OtterAppState;
 import org.hiero.otter.fixtures.internal.AbstractNode;
 import org.hiero.otter.fixtures.internal.result.NodeResultsCollector;
-import org.hiero.otter.fixtures.internal.result.SingleNodeLogResultImpl;
 import org.hiero.otter.fixtures.internal.result.SingleNodeMarkerFileResultImpl;
 import org.hiero.otter.fixtures.internal.result.SingleNodePcesResultImpl;
+import org.hiero.otter.fixtures.logging.internal.InMemorySubscriptionManager;
 import org.hiero.otter.fixtures.result.SingleNodeConsensusResult;
 import org.hiero.otter.fixtures.result.SingleNodeLogResult;
 import org.hiero.otter.fixtures.result.SingleNodeMarkerFileResult;
@@ -126,7 +128,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
         super(selfId, roster);
         logging.addNodeLogging(selfId, outputDirectory);
         try {
-            ThreadContext.put(THREAD_CONTEXT_NODE_ID, this.selfId.toString());
+            ThreadContext.put(THREAD_CONTEXT_NODE_ID, Long.toString(this.selfId.id()));
 
             this.randotron = requireNonNull(randotron);
             this.time = requireNonNull(time);
@@ -149,7 +151,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
     @Override
     public void killImmediately() throws InterruptedException {
         try {
-            ThreadContext.put(THREAD_CONTEXT_NODE_ID, selfId.toString());
+            ThreadContext.put(THREAD_CONTEXT_NODE_ID, Long.toString(this.selfId.id()));
 
             doShutdownNode();
 
@@ -184,7 +186,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
     @Override
     public void start() {
         try {
-            ThreadContext.put(THREAD_CONTEXT_NODE_ID, selfId.toString());
+            ThreadContext.put(THREAD_CONTEXT_NODE_ID, Long.toString(this.selfId.id()));
 
             throwIfIn(RUNNING, "Node has already been started.");
             throwIfIn(DESTROYED, "Node has already been destroyed.");
@@ -211,7 +213,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
     @Override
     public void submitTransaction(@NonNull final byte[] transaction) {
         try {
-            ThreadContext.put(THREAD_CONTEXT_NODE_ID, selfId.toString());
+            ThreadContext.put(THREAD_CONTEXT_NODE_ID, Long.toString(this.selfId.id()));
 
             throwIfIn(INIT, "Node has not been started yet.");
             throwIfIn(SHUTDOWN, "Node has been shut down.");
@@ -240,7 +242,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
     @Override
     @NonNull
     public SingleNodeConsensusResult newConsensusResult() {
-        return resultsCollector.getConsensusResult();
+        return resultsCollector.newConsensusResult();
     }
 
     /**
@@ -249,7 +251,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
     @NonNull
     @Override
     public SingleNodeLogResult newLogResult() {
-        return new SingleNodeLogResultImpl(selfId, Set.of());
+        return resultsCollector.newLogResult();
     }
 
     /**
@@ -258,7 +260,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
     @Override
     @NonNull
     public SingleNodePlatformStatusResult newPlatformStatusResult() {
-        return resultsCollector.getStatusProgression();
+        return resultsCollector.newStatusProgression();
     }
 
     /**
@@ -297,7 +299,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
         if (lifeCycle == RUNNING) {
             assert model != null; // model must be initialized if lifeCycle is STARTED
             try {
-                ThreadContext.put(THREAD_CONTEXT_NODE_ID, selfId.toString());
+                ThreadContext.put(THREAD_CONTEXT_NODE_ID, Long.toString(this.selfId.id()));
                 model.tick();
             } finally {
                 ThreadContext.remove(THREAD_CONTEXT_NODE_ID);
@@ -315,7 +317,7 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
      */
     void destroy() throws InterruptedException {
         try {
-            ThreadContext.put(THREAD_CONTEXT_NODE_ID, selfId.toString());
+            ThreadContext.put(THREAD_CONTEXT_NODE_ID, Long.toString(this.selfId.id()));
 
             resultsCollector.destroy();
             doShutdownNode();
@@ -429,6 +431,13 @@ public class TurtleNode extends AbstractNode implements Node, TurtleTimeManager.
         platformWiring
                 .getStatusStateMachineOutputWire()
                 .solderTo("nodePlatformStatusCollector", "platformStatus", this::handlePlatformStatusChange);
+
+        InMemorySubscriptionManager.INSTANCE.subscribe(logEntry -> {
+            if (Objects.equals(logEntry.nodeId(), selfId)) {
+                resultsCollector.addLogEntry(logEntry);
+            }
+            return lifeCycle == DESTROYED ? UNSUBSCRIBE : CONTINUE;
+        });
 
         platform = platformComponentBuilder.build();
         platformStatus = PlatformStatus.STARTING_UP;
